requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
phonenumberslite==8.13.27
Pillow==10.1.0
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
